% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowpa.R
\name{fill_sinks}
\alias{fill_sinks}
\title{\enumerate{
\item Rellenar sumideros
}}
\usage{
fill_sinks(dem)
}
\arguments{
\item{dem}{un objeto SpatRaster (elevacion)}
}
\value{
un objeto SpatRaster con los sumideros rellenados
}
\description{
\code{fill_sinks} procesa un Modelo Digital de Elevaciones (DEM)
para eliminar imperfecciones y hoyos, evitando que el agua se estanque
}
