% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowpa.R
\name{flow_accumul}
\alias{flow_accumul}
\title{\enumerate{
\item Acumulacion de flujo
}}
\usage{
flow_accumul(dir_flow)
}
\arguments{
\item{dir_flow}{un raster de direccion de flujo, resultado de la funcion 'Direccion de flujo'}
}
\value{
un objeto SpatRaster con el conteo de celdas acumuladas
}
\description{
\code{flow_accumul} calcula cuantas celdas vierten agua en cada celda individual.
}
\examples{
\dontrun{
acu_flow <- flow_accumul(dir_flow)
}
}
