% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowpa.R
\name{flow_direct}
\alias{flow_direct}
\title{\enumerate{
\item Direccion de flujo
}}
\usage{
flow_direct(dem_fill)
}
\arguments{
\item{dem_fill}{un objeto SpatRaster sin sumideros, resultado de la funcion 'Rellenar sumideros'}
}
\value{
un objeto SpatRaster con códigos de direccion
}
\description{
\code{flow_direct} calcula hacia cual de los 8 píxeles vecinos fluira el agua por gravedad
}
