% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowpa.R
\name{rios_vect}
\alias{rios_vect}
\title{\enumerate{
\item Vectorizacion
}}
\usage{
rios_vect(raster_rios)
}
\arguments{
\item{raster_rios}{un raster binario de rios filtrados, resultado de la funcion 'Definir el umbral'}
}
\value{
un objeto sf de tipo multilinestring
}
\description{
\code{rios_vect} convierte los pixeles filtrados en polilineas vectoriales
}
\examples{
\dontrun{
rios_dem <- rios_vect(raster_rios)
}
}
