% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowpa.R
\name{threshold}
\alias{threshold}
\title{\enumerate{
\item Definir el umbral
}}
\usage{
threshold(acu_flow, umbral = 5000)
}
\arguments{
\item{acu_flow}{un raster de acumulacion de flujo, resultado de la funcion 'Acumulacion de flujo'}

\item{umbral}{numerico. Valor minimo para considerar que es un rio (ej. 5000)}
}
\value{
Un objeto SpatRaster binario (1 es rio, NA es no rio)
}
\description{
\code{threshold} filtra el raster para quedarse solo con los pixeles que superen un valor
}
